Fix search field to always follow the value from the URL
Because of `data-turbo-permanent` the search field still had an old value, even if force reloading the page. We therefore have to reset it manually. `data-turbo-permanent` is required so that we can replace the whole page while searching, without recreating the input field and hence moving the cursor to the beginning.

[Closes #198]